+++ op-proposer/proposer/abi_test.go
@@ -103,13 +103,13 @@ func TestManualABIPacking(t *testing.T) {
 
 	output = testutils.RandomOutputResponse(rng)
 
-	txData, err = proposeL2OutputDGFTxData(dgfAbi, uint8(0), output)
+	txData, err = proposeL2OutputDGFTxData(dgfAbi, uint32(0), output)
 	require.NoError(t, err)
 
 	opts.GasLimit = 100_000
 	dgfTx, err := dgf.Create(
 		opts,
-		uint8(0),
+		uint32(0),
 		output.OutputRoot,
 		math.U256Bytes(new(big.Int).SetUint64(output.BlockRef.Number)),
 	)
